+++ src/exchanges/BybitSpot.ts
@@ -49,6 +49,7 @@ class BybitSpot extends BaseExchangeClient {
       apiKey: args.config.apiKey,
       secret: args.config.secret,
       logger: args.logger,
+      httpsAgent: args.config.httpsAgent,
     });
 
     this.publicStream = new BybitPublicStream(